Quick Fix 1
Ok. Those three errors were coming because the functions were not declared. Why were they not declared? Because they will be declared ONLY if the macro CONFIG_USB_GADGET_MSM_72K is defined. Thats why the #ifdef condition has been used. By removing this condition we are making sure that those functions in question are DEFINITELY declared. Therefore by theory,the error should go now if that was the problem.

NOTE: I have analyzed the parent repository and found out that CONFIG_USB_GADGET_MSM_72K is not defined ANYWHERE. Therefore there is a small possibility that these functions might be invalid without it.(EVEN if the compile error goes away)
If that is the case,then we will have to define CONFIG_USB_GADGET_MSM_72K ourselves somewhere in the config. We have to find out to what exactly we have to define it to.

P.S The proposition is purely out of my knowledge in the programming language and having no proper idea over android.