Fix by Vikas
the value is being changed to 100mA,
this means that through USB,the device will ALWAYS receive 100mA. This might be the case even if the charging is completed to 100%,it'll continue to recieve 100mA.
Vikas says that since its very little,it doesnt make a difference. Try the other quick fix,if it doesnt work,this might work.