Quick fix 1p2
Since that was the first error that was shown,the object file would've not been created BECAUSE of that. This happens when the rules set to the compiler is strict and it does not allow you to over ride this warnings anyways and compile like in other popular compilers.
Therefore the quickfix is made to AVOID the warning itself.
the structure variable msm_camera_device_data is not used anywhere in the repository. So it MAY or may not affect the working. Even if it does,once the kernel is booting we can fix the errors. First priority = bootable kernel.