Qfp3
Making the function static will restrict its scope to that file.
That error actually should'nt be coming because in the file in which the error is coming,the header has been included.
Just try this and tell me if it'll work. BTW this fix probably would'nt affect the working in anyway in my opinion. (according to my limited knowledge)